% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\docType{data}
\name{orange}
\alias{orange}
\title{Retro colour palettes}
\format{
A named list of character vectors (hex colours).
}
\usage{
orange
}
\description{
A collection of retro-inspired colour palettes.
}
\keyword{datasets}
