% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scale_fill_retro}
\alias{scale_fill_retro}
\alias{scale_colour_retro}
\title{ggplot2 scale helpers (optional)}
\usage{
scale_fill_retro(palette, ...)

scale_colour_retro(palette, ...)
}
\arguments{
\item{palette}{Name of palette in \code{retro_pals}.}

\item{...}{Passed to ggplot2 scale functions.}
}
\description{
ggplot2 scale helpers (optional)
}
